Hunter tracks food (#50)
* Add estimate to food for a*

* Added double the cost estimate to the subgoal to the cost of a tile

* Added option settings

* Save and load value

close #50

+++ Assets/Scripts/PathFinding/Path.cs
@@ -8,9 +8,9 @@ public class Path<Node>: IEnumerable<Node>
 {
     public Node LastStep { get; private set; }
     public Path<Node> PreviousSteps { get; private set; }
-    public double TotalCost { get; private set; }
+    public int TotalCost { get; private set; }
 
-    private Path(Node lastStep, Path<Node> previousSteps, double totalCost)
+    private Path(Node lastStep, Path<Node> previousSteps, int totalCost)
     {
         LastStep = lastStep;
         PreviousSteps = previousSteps;
@@ -19,7 +19,7 @@ public class Path<Node>: IEnumerable<Node>
 
     public Path(Node start) : this(start, null, 0) { }
 
-    public Path<Node> AddStep(Node step, double stepCost)
+    public Path<Node> AddStep(Node step, int stepCost)
     {
         return new Path<Node>(step, this, TotalCost + stepCost);
     }
